1.2 Added input validation
Added verify() to abstract class Data_Object (which means that it was required in nearly all objects in object model).
The purpose of verify is to check that the objects ("parents") on which an object depends are defined, loaded, and have the correct structure in their underlying data.  verify() is run automatically in load() (which, in turn, is run automatically in ensure() ).
While this is not a bug fix, it does improve error reporting by giving more reasonable error messages for common issues (like a file name has not been defined yet).